finished matching all the functions in ftdonkey.s, needs cleanup, fix some fake stuff, and some of the files are probably ftwalkcommon.c files
