+++ include/functions.h
@@ -2709,12 +2709,12 @@ void func_800DF0D0();
 // ? func_800DF8B0(?);
 // ? func_800DF8E8(?);
 // ? func_800DF910(?);
-// ? func_800DF938(?);
-// ? func_800DF980(?);
-// ? func_800DFA70(?);
-// ? func_800DFA98(?);
-// ? func_800DFAE4(?);
-// ? func_800DFBF8(?);
+// ? ftDonkey_800DF938(?);
+// ? ftDonkey_800DF980(?);
+// ? ftDonkey_800DFA70(?);
+// ? ftDonkey_800DFA98(?);
+// ? ftDonkey_800DFAE4(?);
+// ? ftDonkey_GetWalkType_800DFBF8(?);
 // ? ftDonkey_800DFC70(?);
 // ? ftDonkey_800DFCA4(?);
 // ? ftDonkey_800DFDDC(?);
