--- src/melee/ft/chara/ftDonkey/ftdonkey-4.c
@@ -1,16 +1,0 @@
-#include "ftdonkey.h"
-
-s32 ftDonkey_800E0134(HSD_GObj* fighterObj) {
-    if (func_800C97A8(fighterObj)) {
-        ftDonkey_800E017C(fighterObj);
-        return 1;
-    }
-    return 0;
-}
-
-
-void ftDonkey_800E017C(HSD_GObj* fighterObj) {
-    Fighter* fighter = getFighterPlus(fighterObj);
-    ftDonkeyAttributes* donkey_attr = getFtSpecialAttrs2CC(fighter);
-    func_800C9840(fighterObj, donkey_attr->x0 + 4, 0, 0.0f, donkey_attr->x20, 0.0f);
-}
