+++ src/melee/ft/chara/ftDonkey/ftdonkey2.c
@@ -1,110 +1,37 @@
 #include "ftdonkey.h"
 
-void ftDonkey_OnDeath(HSD_GObj* fighterObj) {
-    Fighter* fighter = fighterObj->user_data;
-    fighter->sa.dk.x222C = 0;
-    func_80074A4C(fighterObj, 0, 0);
-}
 
-void ftDonkey_8010D774(HSD_GObj* fighterObj) {
-    ftDonkey_DestroyAllEffectsPlus(fighterObj);
-    ftDonkey_DestroyAllEffects_SpecialHi(fighterObj);
-}
-
-void ftDonkey_OnItemPickup(HSD_GObj* fighterObj, BOOL bool) {
-    Fighter_OnItemPickup(fighterObj, bool, 1, 1);
-}
-
-void ftDonkey_OnItemInvisible(HSD_GObj* fighterObj) {
-    Fighter_OnItemInvisible(fighterObj, 1);
-}
-
-void ftDonkey_OnItemVisible(HSD_GObj* fighterObj) {
-    Fighter_OnItemVisible(fighterObj, 1);
-}
-
-void ftDonkey_OnItemDrop(HSD_GObj* gobj, BOOL bool1)
-{
-    Fighter_OnItemDrop(gobj, bool1, 1, 1);
-}
-
-void func_8010D96C(HSD_GObj* fighterObj) {
-    Fighter* fighter = fighterObj->user_data;
-    ftDonkeyAttributes* attr = fighter->x2D4_specialAttributes;
-    if (fighter->sa.dk.x222C == attr->x2C) {
-        func_800BFFD0(fighter, 0x39, 0);
+void ftDonkey_800DF9F0(HSD_GObj* fighterObj) {
+    if ((!func_80094EA4(fighterObj)) && (!ftDonkey_800E0378(fighterObj)) && (!ftDonkey_800E0134(fighterObj))) {
+        s32 result = ftDonkey_800DFA98(fighterObj);
+        if (result) { return;  }
     }
 }
 
-
-void ftDonkey_OnLoad(HSD_GObj* fighterObj) {
-    Fighter* fighter = fighterObj->user_data;
-    ftData* ftdata = fighter->x10C_ftData;
-    ftDonkeyAttributes* ftData_attr = ftdata->ext_attr;
-    
-    ftData_attr->x8 = func_8001E8F8(func_80085E50(fighter, 0x128U));
-    ftData_attr->xC = func_8001E8F8(func_80085E50(fighter, 0x129U));
-    ftData_attr->x10 = func_8001E8F8(func_80085E50(fighter, 0x12AU));
-    PUSH_ATTRS(fighter, ftDonkeyAttributes);
-    
-    fighter->x2222_flag.bits.b0 = 1;
-    fighter->x2CC = fighter->x2D4_specialAttributes;
+void ftDonkey_800DFA50(HSD_GObj* fighterObj) {
+    func_80084F3C(fighterObj);
 }
 
-void ftDonkey_LoadSpecialAttrs(HSD_GObj* fighterObj) {
-    COPY_ATTRS(fighterObj, ftDonkeyAttributes);
+
+void ftDonkey_800DFA70(HSD_GObj* fighterObj) {
+    func_8008403C(fighterObj, &ftDonkey_800E0294);
 }
 
-void ftDonkey_OnKnockbackEnter(HSD_GObj* fighterObj) {
-    Fighter_OnKnockbackEnter(fighterObj, 1);
-}
-
-void ftDonkey_OnKnockbackExit(HSD_GObj* fighterObj) {
-    Fighter_OnKnockbackExit(fighterObj, 1);
-}
-
-void ftDonkey_8010DB3C(HSD_GObj* fighterObj) { 
-    s32 unused[2]; /// can't remove with get inlines
-    BOOL bool1;
-    Fighter* fighter = fighterObj->user_data;
-    ftDonkeyAttributes* donkey_attr = fighter->x2D4_specialAttributes;
-    CollData* colldata = &fighter->x6F0_collData; 
-
-    if (fighter->x2210_ThrowFlags.b3) {
-        fighter->x2210_ThrowFlags.b3 = 0;
-        bool1 = 1;
-    } else {
-        bool1 = 0;
+s32 ftDonkey_800DFA98(HSD_GObj* fighterObj) {
+    if (ftDonkey_800DFC70(fighterObj)) {
+        ftDonkey_800DFAE4(fighterObj, 0.0f);
+        return 1;
     }
-    
-    if ((bool1) && ((colldata->x134_envFlags & 0x18000))) {
-        Vec vec_list[4];
+    return 0;
+}
 
-        s32 i;
-        for (i = 0; i < 4; i++) {
-
-            f32 temp_f5 = (donkey_attr->x68 * i) - (donkey_attr->x68 * 1.5f);
-            f32 temp_f3 = donkey_attr->x6C * fighter->x2C_facing_direction;
-            f32 temp_f6 = temp_f5 + temp_f3;
-
-            if (!func_80056C54(
-                    colldata->x14C_groundIndex, 
-                    &fighter->xB0_pos, 
-                    0, 
-                    &vec_list[i], 
-                    0, 
-                    0, 
-                    temp_f6,
-                    donkey_attr->x70, 
-                    donkey_attr->x68 * 1.5f, 
-                    donkey_attr->x68
-                )) {
-
-                vec_list[i] = fighter->xB0_pos;
-            }
-
-            vec_list[i].y += 2.0f; 
-            func_8007B8A8(&fighter->x914[i], &vec_list[i]);
-        }
-    }
+void ftDonkey_800DFAE4(HSD_GObj* fighterObj, f32 argf) {
+    Fighter* fighter = getFighter(fighterObj);
+    ftDonkeyAttributes* donkey_attr = getFtSpecialAttrs2CC(fighter);
+    ftDonkey_800DFCA4(
+        fighterObj, donkey_attr->x0 + 1, 0, 
+        argf, 
+        donkey_attr->x8, donkey_attr->xC,  donkey_attr->x10, donkey_attr->x14, donkey_attr->x18, donkey_attr->x1C, 
+        1.0f
+    ); 
 }
